RX for login is now added to the presenter
